UniC/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
regex==2024.9.11
